- Fix a release datatype warning (or two)
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1950 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/LuaScriptable.cpp
@@ -1011,8 +1011,8 @@ int LuaExports::isPartyInLevelRange(lua_State *luaVm) {
 	Party *p = player->getParty();
 	bool iswithin = false;
 	if (p != 0) {
-		int32_t lowbound = lua_tointeger(luaVm, 1);
-		int32_t highbound = lua_tointeger(luaVm, 2);
+		uint8_t lowbound = lua_tointeger(luaVm, 1);
+		uint8_t highbound = lua_tointeger(luaVm, 2);
 		iswithin = p->isWithinLevelRange(lowbound, highbound);
 	}
 	lua_pushboolean(luaVm, iswithin);
